requirements.txt
Flask==0.10.1
mysql-connector-python==1.1.6